+++ tests/hazmat/primitives/test_hkdf_vectors.py
@@ -21,7 +21,7 @@ from ...utils import load_nist_vectors
 )
 @pytest.mark.requires_backend_interface(interface=HMACBackend)
 class TestHKDFSHA1(object):
-    test_HKDFSHA1 = generate_hkdf_test(
+    test_hkdfsha1 = generate_hkdf_test(
         load_nist_vectors,
         os.path.join("KDF"),
         ["rfc-5869-HKDF-SHA1.txt"],
@@ -35,7 +35,7 @@ class TestHKDFSHA1(object):
 )
 @pytest.mark.requires_backend_interface(interface=HMACBackend)
 class TestHKDFSHA256(object):
-    test_HKDFSHA1 = generate_hkdf_test(
+    test_hkdfsha256 = generate_hkdf_test(
         load_nist_vectors,
         os.path.join("KDF"),
         ["rfc-5869-HKDF-SHA256.txt"],