+++ tests/hazmat/primitives/utils.py
@@ -387,13 +387,13 @@ def kbkdf_counter_mode_test(backend, params):
 
     algorithm = supported_algorithms.get(params.get('prf'))
     if algorithm is None or not backend.hmac_supported(algorithm()):
-        pytest.skip("KBKDF does not support algorithm: {0}".format(
+        pytest.skip("KBKDF does not support algorithm: {}".format(
             params.get('prf')
         ))
 
     ctr_loc = supported_counter_locations.get(params.get("ctrlocation"))
     if ctr_loc is None or not isinstance(ctr_loc, CounterLocation):
-        pytest.skip("Does not support counter location: {0}".format(
+        pytest.skip("Does not support counter location: {}".format(
             params.get('ctrlocation')
         ))